+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 2.0.1 (in development)
 
-- Added `TenantRelationTrait`
+- Added `TenantRelationTrait` and `TenantQueryTrait`